DCON-4774 fix: address Claude Code Review findings (try/finally nit, duplicate suites)
- personMatchManager.test.js: wrap the 5 superagent.post monkey-patch
  test blocks in try/finally so an assertion failure between patch and
  restore can't leak the jest mock into later tests in the same file
  (nock-based tests would then fail with confusing, unrelated errors).

- Removed 10 duplicate test-file pairs (~20 files down to 10) where
  the generation tool guessed two different relative paths for the
  same source module and kept both copies. For each pair, kept the
  file with more test coverage and put it at the path that mirrors the
  source module's actual location (matching every other file in this
  PR); deleted the other. Where the more-complete copy was at the
  wrong path, moved it (adjusting relative require depths) rather than
  keeping the thinner copy just because it happened to be correctly
  located: databaseCursor.test.js, operations/streaming/resourceWriters/
  fhirResourceWriterFactory.test.js, preSaveHandlers/handlers/
  sourceIdColumnHandler.test.js.

Not addressed here (out of scope, tracked separately): the resourceMerger.js
access-tag restoration gap the review surfaced is a pre-existing production
issue, not something this test-only PR introduces or should fix.

Re-verified: 368 files (378 - 10 removed duplicates), all pass
individually and combined; lint clean.

Jira: DCON-4774
