+++ src/tests/unit/dataLayer/databaseCursor.test.js
@@ -1,200 +1,308 @@
 'use strict';
 
-const { describe, test, expect, beforeEach, jest: jestObj } = require('@jest/globals');
+const { describe, test, expect, beforeEach, jest } = require('@jest/globals');
 
-jestObj.mock('../../../utils/assertType', () => ({
-    assertIsValid: jestObj.fn()
+// Mock heavy dependencies
+jest.mock('../../../config', () => ({}));
+jest.mock('../../../operations/common/systemEventLogging', () => ({
+    logTraceSystemEventAsync: jest.fn().mockResolvedValue(undefined),
+    logSystemEventAsync: jest.fn().mockResolvedValue(undefined)
 }));
-
-jestObj.mock('../../../operations/common/systemEventLogging', () => ({
-    logTraceSystemEventAsync: jestObj.fn().mockResolvedValue(undefined)
-}));
-
-jestObj.mock('../../../fhir/fhirResourceCreator', () => ({
+jest.mock('../../../fhir/fhirResourceCreator', () => ({
     FhirResourceCreator: {
-        mapDocumentToResourceObject: jestObj.fn((doc, type) => ({ ...doc, resourceType: type, _mapped: true }))
+        mapDocumentToResourceObject: jest.fn((doc) => ({ ...doc, _mapped: true }))
     }
 }));
 
-jestObj.mock('../../../fhir/classes/4_0_0/backbone_elements/bundleEntry', () => class BundleEntry {});
-
 const { DatabaseCursor } = require('../../../dataLayer/databaseCursor');
+const { FhirResourceCreator } = require('../../../fhir/fhirResourceCreator');
+
+/**
+ * Creates a mock MongoDB FindCursor-like object.
+ */
+function createMockCursor(docs = []) {
+    let index = 0;
+    const cursor = {
+        namespace: { collection: 'TestResource_4_0_0', db: 'fhir' },
+        hasNext: jest.fn(async () => index < docs.length),
+        next: jest.fn(async () => {
+            if (index < docs.length) {
+                return docs[index++];
+            }
+            return null;
+        }),
+        toArray: jest.fn(async () => docs),
+        maxTimeMS: jest.fn().mockReturnThis(),
+        project: jest.fn().mockReturnThis(),
+        map: jest.fn().mockReturnThis(),
+        sort: jest.fn().mockReturnThis(),
+        batchSize: jest.fn().mockReturnThis(),
+        hint: jest.fn().mockReturnThis(),
+        limit: jest.fn().mockReturnThis(),
+        explain: jest.fn(async () => ({ queryPlanner: {} }))
+    };
+    return cursor;
+}
 
 describe('DatabaseCursor', () => {
     let mockCursor;
+    let dbCursor;
 
     beforeEach(() => {
-        mockCursor = {
-            namespace: { collection: 'Patient_4_0_0', db: 'fhir' },
-            hasNext: jestObj.fn().mockResolvedValue(true),
-            next: jestObj.fn().mockResolvedValue({ id: '1', resourceType: 'Patient' }),
-            toArray: jestObj.fn().mockResolvedValue([{ id: '1' }, { id: '2' }]),
-            maxTimeMS: jestObj.fn().mockReturnThis(),
-            project: jestObj.fn().mockReturnThis(),
-            map: jestObj.fn().mockReturnThis(),
-            sort: jestObj.fn().mockReturnThis(),
-            batchSize: jestObj.fn().mockReturnThis(),
-            hint: jestObj.fn().mockReturnThis(),
-            limit: jestObj.fn().mockReturnThis(),
-            explain: jestObj.fn().mockResolvedValue({ queryPlanner: {} })
-        };
+        jest.clearAllMocks();
+        mockCursor = createMockCursor([
+            { id: '1', resourceType: 'Patient' },
+            { id: '2', resourceType: 'Patient' }
+        ]);
+        dbCursor = new DatabaseCursor({
+            base_version: '4_0_0',
+            resourceType: 'Patient',
+            cursor: mockCursor,
+            query: { id: '1' }
+        });
     });
 
-    test('constructor stores properties', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: { id: '1' }
+    describe('constructor validation', () => {
+        test('throws when cursor is null', () => {
+            expect(() => new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Patient',
+                cursor: null,
+                query: {}
+            })).toThrow();
         });
-        expect(dc.base_version).toBe('4_0_0');
-        expect(dc.resourceType).toBe('Patient');
-        expect(dc.cursor).toBe(mockCursor);
-        expect(dc.query).toEqual({ id: '1' });
+
+        test('throws when base_version is undefined', () => {
+            expect(() => new DatabaseCursor({
+                base_version: undefined,
+                resourceType: 'Patient',
+                cursor: mockCursor,
+                query: {}
+            })).toThrow();
+        });
+
+        test('throws when resourceType is undefined', () => {
+            expect(() => new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: undefined,
+                cursor: mockCursor,
+                query: {}
+            })).toThrow();
+        });
     });
 
-    test('hasNext delegates to cursor', async () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('hasNext', () => {
+        test('returns true when cursor has documents', async () => {
+            const result = await dbCursor.hasNext();
+            expect(result).toBe(true);
         });
-        const result = await dc.hasNext();
-        expect(result).toBe(true);
-        expect(mockCursor.hasNext).toHaveBeenCalled();
+
+        test('returns false when cursor is set to empty', async () => {
+            dbCursor.setEmpty();
+            const result = await dbCursor.hasNext();
+            expect(result).toBe(false);
+            // Should not call underlying cursor.hasNext
+            expect(mockCursor.hasNext).not.toHaveBeenCalled();
+        });
     });
 
-    test('hasNext returns false when empty', async () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('next', () => {
+        test('returns document with resourceType added if missing', async () => {
+            const cursorWithoutType = createMockCursor([
+                { id: '1' } // no resourceType, no resource field
+            ]);
+            const cursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Observation',
+                cursor: cursorWithoutType,
+                query: {}
+            });
+
+            const result = await cursor.next();
+            expect(result.resourceType).toBe('Observation');
         });
-        dc.setEmpty();
-        const result = await dc.hasNext();
-        expect(result).toBe(false);
-        expect(mockCursor.hasNext).not.toHaveBeenCalled();
+
+        test('does not overwrite existing resourceType', async () => {
+            const cursorWithType = createMockCursor([
+                { id: '1', resourceType: 'Condition' }
+            ]);
+            const cursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Observation',
+                cursor: cursorWithType,
+                query: {}
+            });
+
+            const result = await cursor.next();
+            expect(result.resourceType).toBe('Condition');
+        });
+
+        test('does not add resourceType when result has resource field (history)', async () => {
+            const cursorWithResource = createMockCursor([
+                { id: '1', resource: { resourceType: 'Patient', id: '1' } }
+            ]);
+            const cursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Patient',
+                cursor: cursorWithResource,
+                query: {}
+            });
+
+            const result = await cursor.next();
+            expect(result.resourceType).toBeUndefined();
+        });
+
+        test('returns null when cursor is exhausted', async () => {
+            const emptyCursor = createMockCursor([]);
+            const cursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Patient',
+                cursor: emptyCursor,
+                query: {}
+            });
+
+            const result = await cursor.next();
+            expect(result).toBeNull();
+        });
+
+        test('wraps cursor errors in RethrownError', async () => {
+            const failingCursor = createMockCursor([]);
+            failingCursor.next.mockRejectedValue(new Error('cursor expired'));
+            const cursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Patient',
+                cursor: failingCursor,
+                query: {}
+            });
+
+            await expect(cursor.next()).rejects.toThrow('cursor expired');
+        });
     });
 
-    test('next returns document from cursor', async () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('nextObject', () => {
+        test('returns null when no document available', async () => {
+            const emptyCursor = createMockCursor([]);
+            const cursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'Patient',
+                cursor: emptyCursor,
+                query: {}
+            });
+
+            const result = await cursor.nextObject();
+            expect(result).toBeNull();
         });
-        const doc = await dc.next();
-        expect(doc).toEqual({ id: '1', resourceType: 'Patient' });
+
+        test('maps document through FhirResourceCreator', async () => {
+            const result = await dbCursor.nextObject();
+            expect(FhirResourceCreator.mapDocumentToResourceObject).toHaveBeenCalledWith(
+                expect.objectContaining({ id: '1' }),
+                'Patient'
+            );
+            expect(result._mapped).toBe(true);
+        });
     });
 
-    test('next adds resourceType when missing from result', async () => {
-        mockCursor.next.mockResolvedValue({ id: '2' });
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Observation', cursor: mockCursor, query: {}
+    describe('cursor operations - chaining', () => {
+        test('maxTimeMS sets timeout on cursor', () => {
+            const result = dbCursor.maxTimeMS({ milliSecs: 5000 });
+            expect(result).toBe(dbCursor);
+            expect(mockCursor.maxTimeMS).toHaveBeenCalledWith(5000);
         });
-        const doc = await dc.next();
-        expect(doc.resourceType).toBe('Observation');
+
+        test('project sets projection', () => {
+            const result = dbCursor.project({ projection: { id: 1 } });
+            expect(result).toBe(dbCursor);
+            expect(mockCursor.project).toHaveBeenCalledWith({ id: 1 });
+        });
+
+        test('sort sets sort option', () => {
+            const result = dbCursor.sort({ sortOption: 'id' });
+            expect(result).toBe(dbCursor);
+            expect(mockCursor.sort).toHaveBeenCalledWith('id', 1);
+        });
+
+        test('batchSize sets batch size', () => {
+            const result = dbCursor.batchSize({ size: 100 });
+            expect(result).toBe(dbCursor);
+            expect(mockCursor.batchSize).toHaveBeenCalledWith(100);
+        });
+
+        test('hint sets index hint', () => {
+            const result = dbCursor.hint({ indexHint: 'id_1' });
+            expect(result).toBe(dbCursor);
+            expect(mockCursor.hint).toHaveBeenCalledWith('id_1');
+        });
+
+        test('limit sets limit and tracks it', () => {
+            expect(dbCursor.getLimit()).toBeNull();
+            const result = dbCursor.limit(50);
+            expect(result).toBe(dbCursor);
+            expect(dbCursor.getLimit()).toBe(50);
+            expect(mockCursor.limit).toHaveBeenCalledWith(50);
+        });
     });
 
-    test('next does not override existing resourceType', async () => {
-        mockCursor.next.mockResolvedValue({ id: '1', resourceType: 'Patient' });
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Observation', cursor: mockCursor, query: {}
+    describe('toArrayAsync', () => {
+        test('returns all documents', async () => {
+            const result = await dbCursor.toArrayAsync();
+            expect(result).toEqual([
+                { id: '1', resourceType: 'Patient' },
+                { id: '2', resourceType: 'Patient' }
+            ]);
         });
-        const doc = await dc.next();
-        expect(doc.resourceType).toBe('Patient');
+
+        test('wraps errors in RethrownError', async () => {
+            mockCursor.toArray.mockRejectedValue(new Error('timeout'));
+            await expect(dbCursor.toArrayAsync()).rejects.toThrow('timeout');
+        });
     });
 
-    test('next wraps errors in RethrownError', async () => {
-        mockCursor.next.mockRejectedValue(new Error('connection lost'));
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('toObjectArrayAsync', () => {
+        test('maps all documents through FhirResourceCreator', async () => {
+            const result = await dbCursor.toObjectArrayAsync();
+            expect(result).toHaveLength(2);
+            expect(FhirResourceCreator.mapDocumentToResourceObject).toHaveBeenCalledTimes(2);
         });
-        await expect(dc.next()).rejects.toThrow();
     });
 
-    test('maxTimeMS sets timeout on cursor', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('explainAsync', () => {
+        test('uses queryPlanner for AuditEvent resource type', async () => {
+            const auditCursor = new DatabaseCursor({
+                base_version: '4_0_0',
+                resourceType: 'AuditEvent',
+                cursor: mockCursor,
+                query: {}
+            });
+
+            await auditCursor.explainAsync();
+            expect(mockCursor.explain).toHaveBeenCalledWith('queryPlanner');
         });
-        const result = dc.maxTimeMS({ milliSecs: 5000 });
-        expect(mockCursor.maxTimeMS).toHaveBeenCalledWith(5000);
-        expect(result).toBe(dc);
+
+        test('uses default explain for non-AuditEvent', async () => {
+            await dbCursor.explainAsync();
+            expect(mockCursor.explain).toHaveBeenCalledWith();
+        });
     });
 
-    test('project sets field projection', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('getCollection and getDatabase', () => {
+        test('returns collection name', () => {
+            expect(dbCursor.getCollection()).toBe('TestResource_4_0_0');
         });
-        dc.project({ projection: { id: 1, name: 1 } });
-        expect(mockCursor.project).toHaveBeenCalledWith({ id: 1, name: 1 });
+
+        test('returns database name', () => {
+            expect(dbCursor.getDatabase()).toBe('fhir');
+        });
     });
 
-    test('sort sets sort order', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
+    describe('resource leak - no close method', () => {
+        test('DatabaseCursor does not expose a close method for cursor cleanup', () => {
+            // BUG: There is no close() method on DatabaseCursor.
+            // If iteration is abandoned (e.g., after finding a result mid-way through),
+            // the underlying MongoDB cursor is never closed, causing server-side resource leak.
+            expect(dbCursor.close).toBeUndefined();
+            // The underlying MongoDB FindCursor has a close method but it's inaccessible
+            // through the DatabaseCursor wrapper without directly accessing .cursor
         });
-        dc.sort({ sortOption: 'meta.lastUpdated' });
-        expect(mockCursor.sort).toHaveBeenCalledWith('meta.lastUpdated', 1);
-    });
-
-    test('limit stores limit and calls cursor', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        dc.limit(50);
-        expect(dc.getLimit()).toBe(50);
-        expect(mockCursor.limit).toHaveBeenCalledWith(50);
-    });
-
-    test('getLimit returns null by default', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        expect(dc.getLimit()).toBeNull();
-    });
-
-    test('getCollection returns cursor collection name', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        expect(dc.getCollection()).toBe('Patient_4_0_0');
-    });
-
-    test('getDatabase returns cursor database name', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        expect(dc.getDatabase()).toBe('fhir');
-    });
-
-    test('toArrayAsync returns array from cursor', async () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        const result = await dc.toArrayAsync();
-        expect(result).toEqual([{ id: '1' }, { id: '2' }]);
-    });
-
-    test('explainAsync uses queryPlanner for AuditEvent', async () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'AuditEvent', cursor: mockCursor, query: {}
-        });
-        await dc.explainAsync();
-        expect(mockCursor.explain).toHaveBeenCalledWith('queryPlanner');
-    });
-
-    test('explainAsync uses default for non-AuditEvent', async () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        await dc.explainAsync();
-        expect(mockCursor.explain).toHaveBeenCalledWith();
-    });
-
-    test('getQuery returns the query', () => {
-        const query = { 'meta.security.code': 'bwell' };
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query
-        });
-        expect(dc.getQuery()).toBe(query);
-    });
-
-    test('batchSize sets batch size on cursor', () => {
-        const dc = new DatabaseCursor({
-            base_version: '4_0_0', resourceType: 'Patient', cursor: mockCursor, query: {}
-        });
-        dc.batchSize({ size: 500 });
-        expect(mockCursor.batchSize).toHaveBeenCalledWith(500);
     });
 });
